♻️ refactor: address swiftlint findings post-impl
Resolve 4 swiftlint --strict violations introduced across PR 3
commits 1-5 (#297, header content unification):

1. **`file_length` on DesignTokens.swift** (414 → 372 lines).
   Move the 4 GameHeader typography tokens (titleScenario / metaRound
   / metaInline / pillStatus) to a new
   `DesignTokens+ExtendedTypography.swift` extension file, mirroring
   the existing `+ExtendedPalette.swift` pattern for color-token
   overflow.

2. **`cyclomatic_complexity` on `SimulationView.phaseDisplayLabel`**
   (11 → 2 + 10).
   Split into `phaseDisplayLabel(for:)` (nil-guard wrapper, complexity
   2) + `phaseName(for:)` (pure switch, complexity 10). Neither half
   crosses the 10-branch ceiling.

3. **`identifier_name` `l3` in DesignTokensTests** (l3 → metaL3).
   Rename inside `headerMetaInkSharesHexWithMetaBaseL3` to satisfy
   the 3-character-minimum rule.

4. **`type_body_length` on DesignTokensTests struct** (288 → ~205
   lines).
   Move §2.12 Header Slots + §3 Typography (GameHeader scale) tests
   to a sibling `DesignTokensTests+GameHeader.swift` extension per
   `.claude/rules/testing.md`. Widen the `approxEqual` helper from
   `private` to internal so the sibling extension can reach it
   (`private` only visible to same-file extensions).

Plus an incidental `Localizable.xcstrings` `Round %lld / %lld` entry
that was lost during the rebase conflict resolution and needed to be
re-added.

No behavioral or API changes — pure structure / naming.

Refs: #297, #273
